+++ webapp/src/test/java/ee/cyber/cdoc2/Cdoc2AuthServerApplicationTest.java
@@ -99,7 +99,7 @@ class Cdoc2AuthServerApplicationTest extends AbstractAuthServerTest {
 
         AuthProcessStatusResponse authStatusResponseBody = performAuthStatusRequest(
             startAuthResponse,
-            2
+            10
         );
 
         assertNotNull(authStatusResponseBody);